authentication/requirements.txt
Django==2.0.1
gunicorn==19.7.1
psycopg2==2.7.3.2
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
requests==2.9.1
